Fixing warning issue related to a conditional testing of multiple elements within the barrier argument
The `if(barrier!="")` testing outputs a warning issue when an object with more than one element is provided, which is often the case for this argument.

I did not test whether this issue would cause a problem in the results, but it did not seem to be a big issue.

In this pull-request, I have replaced `if(barrier!="")` by `if(!identical(barrier, ""))`, which does the test originally intended (whether the barrier object is different than an empty string), without producing a warning related to conditional testing on multiple elements.

The function worked well on both example data tests and on a separate empirical test dataset.